add sourcename to reviews and stuff

+++ 322Api/Models/Device.cs
@@ -8,7 +8,7 @@ namespace _322Api.Models
         [Required]
         public string Name { get; set; }
         public DateTime LastCrawl { get; set; }
-        public int Score { get; set; }
+        public decimal Score { get; set; }
         public string ImageUrl { get; set; }
         public decimal Price { get; set; }
